requirements.txt
awscrt==0.23.5
awsiotsdk==1.22.1
blinker==1.9.0
click==8.1.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==2.1.0
pluggy==1.5.0
pytest==8.3.4
Werkzeug==3.1.3
